big refactoring
moved files around so that it makes more sense
deleted a bunch of useless files
added todo to remind myself of what i still need to do
