+++ packages/open-next/src/adapters/config/util.ts
@@ -1,12 +1,12 @@
 import fs from "fs";
 import path from "path";
-
-import { PublicFiles } from "../../build";
 import {
   NextConfig,
   PrerenderManifest,
   RoutesManifest,
-} from "../types/next-types";
+} from "types/next-types";
+
+import { PublicFiles } from "../../build";
 
 export function loadConfig(nextDir: string) {
   const filePath = path.join(nextDir, "required-server-files.json");
