+++ packages/open-next/src/adapters/plugins/13.5/serverHandler.ts
@@ -1,12 +1,11 @@
 /*eslint-disable simple-import-sort/imports */
-import type { Options, PluginHandler } from "../../types/next-types.js";
-import type { IncomingMessage } from "../../http/request.js";
+import type { Options, PluginHandler } from "types/next-types.js";
+import type { IncomingMessage, OpenNextNodeResponse } from "http/index.js";
 //#override imports
 //@ts-ignore
 import { requestHandler } from "./util.js";
 //@ts-ignore
 import { proxyRequest } from "./routing/util.js";
-import { OpenNextNodeResponse } from "../../http/openNextResponse.js";
 //#endOverride
 
 //#override handler
