+++ packages/open-next/src/converters/aws-apigw-v2.ts
@@ -1,9 +1,9 @@
 import { APIGatewayProxyEventV2, APIGatewayProxyResultV2 } from "aws-lambda";
+import { parseCookies } from "http/util";
+import type { Converter, InternalEvent, InternalResult } from "types/open-next";
 
-import { InternalEvent, InternalResult } from "../adapters/event-mapper";
 import { debug } from "../adapters/logger";
-import { Converter } from "../adapters/types/open-next";
-import { parseCookies } from "../adapters/util";
+import { convertToQuery } from "../core/routing/util";
 import { removeUndefinedFromQuery } from "./utils";
 
 function normalizeAPIGatewayProxyEventV2Body(
@@ -51,7 +51,7 @@ async function convertFromAPIGatewayProxyEventV2(
     body: normalizeAPIGatewayProxyEventV2Body(event),
     headers: normalizeAPIGatewayProxyEventV2Headers(event),
     remoteAddress: requestContext.http.sourceIp,
-    query: removeUndefinedFromQuery(event.queryStringParameters ?? {}),
+    query: removeUndefinedFromQuery(convertToQuery(rawQueryString)),
     cookies:
       event.cookies?.reduce((acc, cur) => {
         const [key, value] = cur.split("=");
