+++ packages/open-next/src/types/aws-lambda.ts
