+++ packages/open-next/tsconfig.json
@@ -5,6 +5,11 @@
     "module": "esnext",
     "lib": ["DOM", "ESNext"],
     "outDir": "./dist",
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    "paths": {
+      "types/*": ["./src/types/*"],
+      "config/*": ["./src/adapters/config/*"],
+      "http/*": ["./src/http/*"],
+    }
   }
 }
